fixing netclient join command

+++ netclient/local/local.go
@@ -136,7 +136,6 @@ OnCalendar=*:*:0/30
 WantedBy=timers.target
 `
 
-
 	servicebytes := []byte(systemservice)
 	timerbytes := []byte(systemtimer)
 
@@ -255,9 +254,13 @@ func RemoveSystemDServices(network string) error {
                 fmt.Println(err)
         }
 	if fullremove {
-	err = os.Remove("/etc/systemd/system/netclient@.service")
+		if FileExists("/etc/systemd/system/netclient@.service") {
+			err = os.Remove("/etc/systemd/system/netclient@.service")
+		}
 	}
-	err = os.Remove("/etc/systemd/system/netclient-"+network+".timer")
+	if FileExists("/etc/systemd/system/netclient-"+network+".timer") {
+		err = os.Remove("/etc/systemd/system/netclient-"+network+".timer")
+	}
 	if err != nil {
                 fmt.Println("Error removing file. Please investigate.")
                 fmt.Println(err)
@@ -286,9 +289,15 @@ func WipeLocal(network string) error{
 
         //home, err := homedir.Dir()
         home := "/etc/netclient"
-        _ = os.Remove(home + "/netconfig-" + network)
-        _ = os.Remove(home + "/nettoken-" + network)
-        _ = os.Remove(home + "/wgkey-" + network)
+	if FileExists(home + "/netconfig-" + network) {
+	        _ = os.Remove(home + "/netconfig-" + network)
+	}
+	if FileExists(home + "/nettoken-" + network) {
+		_ = os.Remove(home + "/nettoken-" + network)
+	}
+	if FileExists(home + "/wgkey-" + network) {
+		_ = os.Remove(home + "/wgkey-" + network)
+	}
 
         ipExec, err := exec.LookPath("ip")
